Partly implemented clean architecture, bug fixes.
