+++ src/presentation/assets/robot.png
