+++ src/presentation/styles/UploadPage.scss
@@ -70,7 +70,6 @@
         flex-direction: column; 
         h3{
           font-size:40vw;
-          position:absolute;
           top:50%;
           transform:translateY(-50%);
           opacity:.1;
